fix: separate integration and unit tests in config, add fast-fail build tags
